implemented replace mode

+++ src/controller/command/mutate_command/replace_command.cc
@@ -1,24 +1,24 @@
-#include "model/ptr_cursor.h"
 #include "controller/controller.h"
+#include "controller/controller.h"
+#include "model/file_buffer.h"
 #include "replace_command.h"
 
 
 namespace VM {
     inline void ReplaceCommand::commandHelper(Controller &controller) const {
-        controller.getBuffer().delete_forward(1);
+        if(deleted) controller.getBuffer().delete_forward(1);
         controller.getBuffer().type(toMutate);
-        controller.getBuffer().ptrCursor.moveLeft();
     }
 
     void ReplaceCommand::doTheCommand(Controller &controller) {
-        deleted = *controller.getBuffer().ptrCursor;
+        if(!controller.getBuffer().ptrCursor.isEOL())
+            deleted = *controller.getBuffer().ptrCursor;
         commandHelper(controller);
     }
 
     void ReplaceCommand::undoTheCommand(Controller &controller) const {
-        controller.getBuffer().delete_forward(1);
-        controller.getBuffer().type(deleted);
-        controller.getBuffer().ptrCursor.moveLeft();
+        controller.getBuffer().delete_backward(1);
+        if(deleted) controller.getBuffer().type(deleted);
     }
 
     void ReplaceCommand::redoTheCommand(Controller &controller) const {
